+++ tests/test_controls.py
@@ -203,96 +203,3 @@ def test_default_sensor_contorls(action, expected):
     for k, v in state.sensor_states.items():
         assert k in new_state.sensor_states
         _check_state_expected(v, new_state.sensor_states[k], expected)
-
-
-@pytest.mark.parametrize(
-    "noise_multiplier,robot,controller",
-    itertools.product(
-        [1.0, 0.0], ["LoCoBot", "LoCoBot-Lite"], ["ILQR", "Proportional", "Movebase"]
-    ),
-)
-def test_pyrobot_noisy_actions(noise_multiplier, robot, controller):
-    scene_graph = hsim.SceneGraph()
-    agent_config = habitat_sim.AgentConfiguration()
-    agent_config.action_space = dict(
-        noisy_move_backward=habitat_sim.ActionSpec(
-            "pyrobot_noisy_move_backward",
-            habitat_sim.PyRobotNoisyActuationSpec(
-                amount=0.25,
-                robot=robot,
-                controller=controller,
-                noise_multiplier=noise_multiplier,
-            ),
-        ),
-        noisy_move_forward=habitat_sim.ActionSpec(
-            "pyrobot_noisy_move_forward",
-            habitat_sim.PyRobotNoisyActuationSpec(
-                amount=0.25,
-                robot=robot,
-                controller=controller,
-                noise_multiplier=noise_multiplier,
-            ),
-        ),
-        noisy_turn_left=habitat_sim.ActionSpec(
-            "pyrobot_noisy_turn_left",
-            habitat_sim.PyRobotNoisyActuationSpec(
-                amount=10.0,
-                robot=robot,
-                controller=controller,
-                noise_multiplier=noise_multiplier,
-            ),
-        ),
-        noisy_turn_right=habitat_sim.ActionSpec(
-            "pyrobot_noisy_turn_right",
-            habitat_sim.PyRobotNoisyActuationSpec(
-                amount=10.0,
-                robot=robot,
-                controller=controller,
-                noise_multiplier=noise_multiplier,
-            ),
-        ),
-        move_backward=habitat_sim.ActionSpec(
-            "move_backward", habitat_sim.ActuationSpec(amount=0.25)
-        ),
-        move_forward=habitat_sim.ActionSpec(
-            "move_forward", habitat_sim.ActuationSpec(amount=0.25)
-        ),
-        turn_left=habitat_sim.ActionSpec(
-            "turn_left", habitat_sim.ActuationSpec(amount=10.0)
-        ),
-        turn_right=habitat_sim.ActionSpec(
-            "turn_right", habitat_sim.ActuationSpec(amount=10.0)
-        ),
-    )
-    agent = habitat_sim.Agent(agent_config)
-    agent.attach(scene_graph.get_root_node().create_child())
-
-    for action in agent_config.action_space:
-        base_action = action.replace("noisy_", "")
-
-        state = agent.state
-        agent.act(base_action)
-        base_state = agent.state
-
-        agent.state = state
-
-        agent.act(f"noisy_{base_action}")
-        noisy_state = agent.state
-
-        with (
-            pytest.raises(AssertionError)
-            if noise_multiplier > 0.0
-            else contextlib.suppress()
-        ):
-            _check_state_same(base_state, noisy_state)
-
-        for k, v in base_state.sensor_states.items():
-            assert k in noisy_state.sensor_states
-            with (
-                pytest.raises(AssertionError)
-                if noise_multiplier > 0.0
-                else contextlib.suppress()
-            ):
-                _check_state_same(v, noisy_state.sensor_states[k])
-
-        agent.state = state
